move config_completion to common
